third_party/blink/web_tests/html/menu/tentative/menu-elements.html

<!DOCTYPE html>
<script src='../../../resources/testharness.js'></script>
<script src='../../../resources/testharnessreport.js'></script>

<menubar>
  <menuitem id="menubaritem">Command 1</menuitem>
  <menuitem>Command 2</menuitem>
</menubar>

<menulist>
 <menuitem id="menulistitem" disabled>Command 1</menuitem>
 <menuitem>Command 2</menuitem>
</menulist>

<script>

test(() => {
 const menubar = document.querySelector("menubar");
 const menubaritem = document.getElementById("menubaritem");

 assert_equals(menubar.constructor, HTMLMenuBarElement);
 assert_equals(menubaritem.constructor, HTMLMenuItemElement);
 assert_false(menubaritem.disabled);
 menubaritem.disabled = true;
 assert_true(menubaritem.disabled);

 const menulist = document.querySelector("menulist");
 const menulistitem = document.getElementById("menulistitem");

 assert_equals(menulist.constructor, HTMLMenuListElement);
 assert_equals(menulistitem.constructor, HTMLMenuItemElement);
 assert_true(menulistitem.disabled);
 menulistitem.disabled = false;
 assert_false(menulistitem.disabled);
}, "Menu elements are HTML elements.");

test(() => {
  const menubaritem = document.getElementById("menubaritem");

  // `defaultChecked`.
  assert_equals(menubaritem.defaultChecked, undefined,
      "menuitem does not have a defaultChecked attribute");

  // `checked` content & IDL attribute behavior.
  // By default, attribute missing/null, which translates to the false IDL
  // boolean state.
  assert_equals(menubaritem.getAttribute('checked'), null);
  assert_false(menubaritem.checked);

  // The IDL attribute reads from the same state that the content attribute
  // writes to.
  menubaritem.setAttribute("checked", "true");
  assert_true(menubaritem.checked);

  // But the IDL attribute does not reflect [1] the content attribute value,
  // which means when it is updated, the content attribute does not change.
  //
  // [1]: https://html.spec.whatwg.org/C#reflect
  menubaritem.checked = false;
  assert_false(menubaritem.checked);
  menubaritem.setAttribute("checked", "true");
}, "menuitem checked attribute");
</script>
